+++ ProjectTalentOnboardingTask.Server/Controllers/ProductController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using ProjectTalentOnboardingTask.Server.Models;
 using Microsoft.EntityFrameworkCore;
+using ProjectTalentOnboardingTask.Server.Models;
 
 namespace ProjectTalentOnboardingTask.Server.Controllers
 {
@@ -8,58 +8,95 @@ namespace ProjectTalentOnboardingTask.Server.Controllers
     [ApiController]
     public class ProductController : ControllerBase
     {
-        private readonly ApplicationDbContext _context;
+        private readonly DBContext _context;
 
-        public ProductController(ApplicationDbContext context)
+        public ProductController(DBContext context)
         {
             _context = context;
         }
 
         // GET: api/<ProductController>
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Product>>> GetProducts()
+        public async Task<ActionResult<IEnumerable<Product>>> GetProductsAsync()
         {
-            return await _context.Product.ToArrayAsync();
+            try
+            {
+                return await _context.Product.ToArrayAsync();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in products fetching {ex.Message}");
+            }
+
         }
 
 
         // GET api/<ProductController>/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Product>> GetProduct(int id)
+        public async Task<ActionResult<Product>> GetProductAsync(int id)
         {
-            var product = await _context.Product.FindAsync(id);
-
-            if (product == null)
+            if (!HasValidId(id))
             {
-                return NotFound();
+                return BadRequest("Invalid ID");
             }
 
-            return product;
+            try
+            {
+                var product = await _context.Product.FindAsync(id);
+
+                if (product == null)
+                {
+                    return NotFound();
+                }
+
+                return product;
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in product fetching {ex.Message}");
+            }
+
+
+
         }
 
         // POST api/<ProductController>
         [HttpPost]
-        public async Task<ActionResult<Product>> PostProduct(Product product)
+        public async Task<ActionResult<Product>> PostProductAsync(Product product)
         {
-            _context.Product.Add(product);
-            await _context.SaveChangesAsync();
+            try
+            {
+                _context.Product.Add(product);
+                await _context.SaveChangesAsync();
 
-            return CreatedAtAction(nameof(GetProduct), new { id = product.Id }, product);
+                return product;
+
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in product saving {ex.Message}");
+            }
+
         }
 
         // PUT api/<ProductController>/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutProduct(int id, Product product)
+        public async Task<IActionResult> PutProductAsync(int id, Product product)
         {
-            if (id != product.Id)
+            if (!HasValidId(id))
             {
-                return BadRequest();
+                return BadRequest("Invalid ID");
             }
-
-            _context.Entry(product).State = EntityState.Modified;
 
             try
             {
+                if (id != product.Id)
+                {
+                    return BadRequest();
+                }
+
+                _context.Entry(product).State = EntityState.Modified;
+
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
@@ -80,24 +117,45 @@ namespace ProjectTalentOnboardingTask.Server.Controllers
         // DELETE api/<ProductController>/5
         [HttpDelete("{id}")]
 
-        public async Task<IActionResult> DeleteProduct(int id)
+        public async Task<IActionResult> DeleteProductAsync(int id)
         {
-            var product = await _context.Product.FindAsync(id);
-            if (product == null)
+            if (!HasValidId(id))
             {
-                return NotFound();
+                return BadRequest("Invalid ID");
             }
 
-            _context.Product.Remove(product);
-            await _context.SaveChangesAsync();
+            try
+            {
+                var product = await _context.Product.FindAsync(id);
+                if (product == null)
+                {
+                    return NotFound();
+                }
 
-            return NoContent();
+                _context.Product.Remove(product);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
+
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in product deleting {ex.Message}");
+            }
+
+
+
         }
 
         private bool ProductExists(int id)
         {
             return _context.Product.Any(e => e.Id == id);
         }
+
+        private bool HasValidId(int id)
+        {
+            return id > 0;
+        }
+
     }
 }
-
