+++ ProjectTalentOnboardingTask.Server/Controllers/StoreController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using ProjectTalentOnboardingTask.Server.Models;
 using Microsoft.EntityFrameworkCore;
+using ProjectTalentOnboardingTask.Server.Models;
 
 namespace ProjectTalentOnboardingTask.Server.Controllers
 {
@@ -8,58 +8,94 @@ namespace ProjectTalentOnboardingTask.Server.Controllers
     [ApiController]
     public class StoreController : ControllerBase
     {
-        private readonly ApplicationDbContext _context;
+        private readonly DBContext _context;
 
-        public StoreController(ApplicationDbContext context)
+        public StoreController(DBContext context)
         {
             _context = context;
         }
 
         // GET: api/<StoreController>
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Store>>> GetStores()
+        public async Task<ActionResult<IEnumerable<Store>>> GetStoresAsync()
         {
-            return await _context.Store.ToArrayAsync();
-        }
+            try
+            {
+                return await _context.Store.ToArrayAsync();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in stores fetching {ex.Message}");
+            }
 
+        }
 
         // GET api/<StoreController>/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Store>> GetStore(int id)
+        public async Task<ActionResult<Store>> GetStoreAsync(int id)
         {
-            var store = await _context.Store.FindAsync(id);
-
-            if (store == null)
+            if (!HasValidId(id))
             {
-                return NotFound();
+                return BadRequest("Invalid ID");
             }
 
-            return store;
+            try
+            {
+                var store = await _context.Store.FindAsync(id);
+
+                if (store == null)
+                {
+                    return NotFound();
+                }
+
+                return store;
+
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in store fetching {ex.Message}");
+            }
+
+
         }
 
         // POST api/<StoreController>
         [HttpPost]
-        public async Task<ActionResult<Store>> PostStore(Store store)
+        public async Task<ActionResult<Store>> PostStoreAsync(Store store)
         {
-            _context.Store.Add(store);
-            await _context.SaveChangesAsync();
+            try
+            {
+                _context.Store.Add(store);
+                await _context.SaveChangesAsync();
 
-            return CreatedAtAction(nameof(GetStore), new { id = store.Id }, store);
+                return store;
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in store saving {ex.Message}");
+            }
+
+
         }
 
         // PUT api/<StoreController>/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutStore(int id, Store store)
+        public async Task<IActionResult> PutStoreAsync(int id, Store store)
         {
-            if (id != store.Id)
+            if (!HasValidId(id))
             {
-                return BadRequest();
+                return BadRequest("Invalid ID");
             }
-
-            _context.Entry(store).State = EntityState.Modified;
 
             try
             {
+                if (id != store.Id)
+                {
+                    return BadRequest();
+                }
+
+                _context.Entry(store).State = EntityState.Modified;
+
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
@@ -80,23 +116,44 @@ namespace ProjectTalentOnboardingTask.Server.Controllers
         // DELETE api/<StoreController>/5
         [HttpDelete("{id}")]
 
-        public async Task<IActionResult> DeleteStore(int id)
+        public async Task<IActionResult> DeleteStoreAsync(int id)
         {
-            var store = await _context.Store.FindAsync(id);
-            if (store == null)
+            if (!HasValidId(id))
             {
-                return NotFound();
+                return BadRequest("Invalid ID");
             }
 
-            _context.Store.Remove(store);
-            await _context.SaveChangesAsync();
+            try
+            {
+                var store = await _context.Store.FindAsync(id);
 
-            return NoContent();
+                if (store == null)
+                {
+                    return NotFound();
+                }
+
+                _context.Store.Remove(store);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
+
+            }
+            catch (Exception ex)
+            {
+                return BadRequest($"error in store deleting {ex.Message}");
+            }
+
+
         }
 
         private bool StoreExists(int id)
         {
             return _context.Store.Any(e => e.Id == id);
+        }
+
+        private bool HasValidId(int id)
+        {
+            return id > 0;
         }
     }
 }
